OZP-192 Rework validation
Domain entities share a "validate" method which will return a list of ValidationError's. The ValidationErrors describe which properties of the Entity are invalid and why in a human readable message.

Provided a lightweight, annotation-based bean validation system to reuse validation logic succinctly. Right now only the @NotBlank annotation is implemented but others could include @Range and @Email

Value Objects, which are immutable and created by the Entity which owns them, must be created in a valid state so these throw exceptions in their constructors.

Updated tests and provided new tests for validation system